oor: delete the legacy global actor and signing-effect actor
In this commit, we remove the old single global OOR actor and the
signing-effect actor now that the per-session registry owns the OOR
service key. The shared symbols the new code still needs -- the
OutboxHandler interface, the metadata-filter alias, and the actor
message codec -- move to shared.go; everything else in actor.go and all
of signing_effect_actor.go is deleted, along with the whole-map
checkpoint codec that backed the old per-mutation blob write.

The old actor's unit tests are deleted with it. The two helpers a few
surviving tests still use, testRetryTransferInputs and
buildIncomingResolveResponse, move to kept test files.

The two SendOOR idempotency RPC tests are converted to drive the
registry instead of the global actor. They now build a per-session
signer, a registry control-plane store on the same database as the
delivery store, and a no-op serverconn ref, so the create-then-dedup
and unlock-on-existing paths run through the real per-session actor end
to end. Both pass. The systest's vestigial OOR actor construction --
created and stopped but never used, since those tests drive the
incoming FSM directly -- is removed.
