+++ oor/actor_durable_message.go
@@ -36,10 +36,6 @@ const (
 )
 
 const (
-	idempotencyKeyPayloadRecordType tlv.Type = 1
-)
-
-const (
 	// listSessionsDirectionRecordType stores the session direction filter.
 	listSessionsDirectionRecordType tlv.Type = 1
 
@@ -864,14 +860,6 @@ func encodeTransferInputSnapshots(inputs []*TransferInputSnapshot) ([]byte,
 	return encodeLengthPrefixedBlobList(blobs)
 }
 
-func decodeTransferInputSnapshots(raw []byte) ([]*TransferInputSnapshot,
-	error) {
-
-	return decodeTransferInputSnapshotsWithLimits(
-		raw, ReceiveLimits{},
-	)
-}
-
 // decodeTransferInputSnapshotsWithLimits decodes transfer-input snapshots
 // using the supplied receive limits for the outer blob list.
 func decodeTransferInputSnapshotsWithLimits(raw []byte,
@@ -1466,48 +1454,6 @@ func decodeSessionPayload(raw []byte) (SessionID, error) {
 	}
 
 	return parseSessionID(sessionBytes)
-}
-
-func encodeIdempotencyKeyPayload(idempotencyKey string) ([]byte, error) {
-	idKey := []byte(idempotencyKey)
-	records := []tlv.Record{
-		tlv.MakePrimitiveRecord(
-			idempotencyKeyPayloadRecordType, &idKey,
-		),
-	}
-
-	stream, err := tlv.NewStream(records...)
-	if err != nil {
-		return nil, err
-	}
-
-	var buf bytes.Buffer
-	if err := stream.Encode(&buf); err != nil {
-		return nil, err
-	}
-
-	return buf.Bytes(), nil
-}
-
-func decodeIdempotencyKeyPayload(raw []byte) (string, error) {
-	var idKey []byte
-	records := []tlv.Record{
-		tlv.MakePrimitiveRecord(
-			idempotencyKeyPayloadRecordType, &idKey,
-		),
-	}
-
-	stream, err := tlv.NewStream(records...)
-	if err != nil {
-		return "", err
-	}
-
-	reader := bytes.NewReader(raw)
-	if _, err := stream.DecodeWithParsedTypes(reader); err != nil {
-		return "", err
-	}
-
-	return string(idKey), nil
 }
 
 // encodeListSessionsPayload encodes the durable list-sessions query filters.
