+++ oor/outgoing_snapshot_codec.go
@@ -11,12 +11,6 @@ import (
 )
 
 const (
-	checkpointVersionRecordType           tlv.Type = 1
-	checkpointSnapshotsRecordType         tlv.Type = 3
-	checkpointIncomingSnapshotsRecordType tlv.Type = 5
-)
-
-const (
 	snapshotVersionRecordType         tlv.Type = 1
 	snapshotSessionIDRecordType       tlv.Type = 3
 	snapshotPhaseRecordType           tlv.Type = 5
@@ -30,189 +24,6 @@ const (
 	// OOR send idempotency key.
 	snapshotIdempotencyKeyRecordType tlv.Type = 21
 )
-
-type sessionsCheckpoint struct {
-	Version           int
-	OutgoingSnapshots []*OutgoingSnapshot
-	IncomingSnapshots []*IncomingSnapshot
-}
-
-func encodeSessionsCheckpoint(checkpoint sessionsCheckpoint) ([]byte, error) {
-	outgoingBlobs := make([][]byte, 0, len(checkpoint.OutgoingSnapshots))
-	for i := range checkpoint.OutgoingSnapshots {
-		raw, err := encodeOutgoingSnapshot(
-			checkpoint.OutgoingSnapshots[i],
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		outgoingBlobs = append(outgoingBlobs, raw)
-	}
-
-	outgoingRaw, err := encodeLengthPrefixedBlobList(outgoingBlobs)
-	if err != nil {
-		return nil, err
-	}
-
-	incomingBlobs := make([][]byte, 0, len(checkpoint.IncomingSnapshots))
-	for i := range checkpoint.IncomingSnapshots {
-		raw, err := encodeIncomingSnapshot(
-			checkpoint.IncomingSnapshots[i],
-		)
-		if err != nil {
-			return nil, err
-		}
-
-		incomingBlobs = append(incomingBlobs, raw)
-	}
-
-	incomingRaw, err := encodeLengthPrefixedBlobList(incomingBlobs)
-	if err != nil {
-		return nil, err
-	}
-
-	version := uint64(checkpoint.Version)
-	records := []tlv.Record{
-		tlv.MakePrimitiveRecord(checkpointVersionRecordType, &version),
-		tlv.MakePrimitiveRecord(
-			checkpointSnapshotsRecordType, &outgoingRaw,
-		),
-		tlv.MakePrimitiveRecord(
-			checkpointIncomingSnapshotsRecordType, &incomingRaw,
-		),
-	}
-
-	stream, err := tlv.NewStream(records...)
-	if err != nil {
-		return nil, err
-	}
-
-	var buf bytes.Buffer
-	if err := stream.Encode(&buf); err != nil {
-		return nil, err
-	}
-
-	return buf.Bytes(), nil
-}
-
-func decodeSessionsCheckpoint(raw []byte) (sessionsCheckpoint, error) {
-	return decodeSessionsCheckpointWithLimits(
-		raw, ReceiveLimits{},
-	)
-}
-
-// decodeSessionsCheckpointWithLimits decodes a checkpoint and applies receive
-// limits to all nested session snapshot lists.
-func decodeSessionsCheckpointWithLimits(raw []byte,
-	limits ReceiveLimits) (sessionsCheckpoint, error) {
-
-	var (
-		version     uint64
-		outgoingRaw []byte
-		incomingRaw []byte
-	)
-
-	records := []tlv.Record{
-		tlv.MakePrimitiveRecord(checkpointVersionRecordType, &version),
-		tlv.MakePrimitiveRecord(
-			checkpointSnapshotsRecordType, &outgoingRaw,
-		),
-		tlv.MakePrimitiveRecord(
-			checkpointIncomingSnapshotsRecordType, &incomingRaw,
-		),
-	}
-
-	stream, err := tlv.NewStream(records...)
-	if err != nil {
-		return sessionsCheckpoint{}, err
-	}
-
-	reader := bytes.NewReader(raw)
-	if _, err := stream.DecodeWithParsedTypes(reader); err != nil {
-		return sessionsCheckpoint{}, err
-	}
-
-	outgoingBlobs, err := decodeLengthPrefixedBlobListWithLimits(
-		outgoingRaw, limits,
-	)
-	if err != nil {
-		return sessionsCheckpoint{}, err
-	}
-
-	outgoingSnapshots := make([]*OutgoingSnapshot, 0,
-		len(outgoingBlobs))
-	for i := range outgoingBlobs {
-		snapshot, err := decodeOutgoingSnapshotWithLimits(
-			outgoingBlobs[i], limits,
-		)
-		if err != nil {
-			return sessionsCheckpoint{}, err
-		}
-
-		outgoingSnapshots = append(outgoingSnapshots, snapshot)
-	}
-
-	incomingSnapshots := make([]*IncomingSnapshot, 0)
-	if len(incomingRaw) != 0 {
-		incomingBlobs, err := decodeLengthPrefixedBlobListWithLimits(
-			incomingRaw, limits,
-		)
-		if err != nil {
-			return sessionsCheckpoint{}, err
-		}
-
-		incomingSnapshots = make(
-			[]*IncomingSnapshot, 0, len(incomingBlobs),
-		)
-		for i := range incomingBlobs {
-			snapshot, err := decodeIncomingSnapshotWithLimits(
-				incomingBlobs[i], limits,
-			)
-			if err != nil {
-				return sessionsCheckpoint{}, err
-			}
-
-			incomingSnapshots = append(
-				incomingSnapshots, snapshot,
-			)
-		}
-	}
-
-	decodedVersion, err := decodeUint64ToInt(version, "checkpoint version")
-	if err != nil {
-		return sessionsCheckpoint{}, err
-	}
-
-	return sessionsCheckpoint{
-		Version:           decodedVersion,
-		OutgoingSnapshots: outgoingSnapshots,
-		IncomingSnapshots: incomingSnapshots,
-	}, nil
-}
-
-func encodeOutgoingSessionsCheckpoint(checkpoint outgoingSessionsCheckpoint) (
-	[]byte, error) {
-
-	return encodeSessionsCheckpoint(sessionsCheckpoint{
-		Version:           checkpoint.Version,
-		OutgoingSnapshots: checkpoint.Snapshots,
-	})
-}
-
-func decodeOutgoingSessionsCheckpoint(raw []byte) (outgoingSessionsCheckpoint,
-	error) {
-
-	checkpoint, err := decodeSessionsCheckpoint(raw)
-	if err != nil {
-		return outgoingSessionsCheckpoint{}, err
-	}
-
-	return outgoingSessionsCheckpoint{
-		Version:   checkpoint.Version,
-		Snapshots: checkpoint.OutgoingSnapshots,
-	}, nil
-}
 
 func encodeOutgoingSnapshot(snapshot *OutgoingSnapshot) ([]byte, error) {
 	if snapshot == nil {
@@ -402,15 +213,6 @@ func decodeUint64ToUint8(value uint64, field string) (uint8, error) {
 	}
 
 	return uint8(value), nil
-}
-
-func decodeUint64ToInt(value uint64, field string) (int, error) {
-	maxInt := uint64(^uint(0) >> 1)
-	if value > maxInt {
-		return 0, fmt.Errorf("%s overflows int: %d", field, value)
-	}
-
-	return int(value), nil
 }
 
 func decodeUint64ToDuration(value uint64, field string) (time.Duration, error) {
